Backend/requirements.txt
asgiref==3.2.10
attrs==20.1.0
certifi==2020.6.20
chardet==3.0.4
defusedxml==0.6.0
Django==3.1
django-allauth==0.42.0
django-oauth-toolkit==1.3.2
django-rest-auth==0.9.5
djangorestframework==3.11.1
djangorestframework-jwt==1.11.0
gunicorn==20.0.4
idna==2.10
importlib-metadata==1.7.0
iniconfig==1.0.1
more-itertools==8.4.0
oauthlib==3.1.0
packaging==20.4
pkg-resources==0.0.0
pluggy==0.13.1
psycopg2-binary==2.8.5
py==1.9.0
PyJWT==1.7.1
pyparsing==2.4.7
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
rest-condition==1.0.3
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
urllib3==1.25.10
zipp==3.1.0